pyproject.toml
[project]
name = "news-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.135.3",
    "google-genai>=1.70.0",
    "groq>=1.1.2",
    "gtts>=2.5.4",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "uvicorn>=0.44.0",
]
